requirements.txt
# requirements.txt
torch
torchvision
transformers
Pillow
matplotlib
numpy
fastapi
uvicorn[standard]
requests
python-multipart
gradio
gradio_image_annotation
segment-anything
